+++ src/utils/import.ts
@@ -12,7 +12,7 @@ import {
   modelOptions,
   _defaultChatConfig,
 } from '@constants/chat';
-import { ExportV1 } from '@type/export';
+import { ExportV1, OpenAIChat } from '@type/export';
 
 export const validateAndFixChats = (chats: any): chats is ChatInterface[] => {
   if (!Array.isArray(chats)) return false;
@@ -87,4 +87,46 @@ export const validateFolders = (
 
 export const validateExportV1 = (data: ExportV1): data is ExportV1 => {
   return validateAndFixChats(data.chats) && validateFolders(data.folders);
+};
+
+// Convert OpenAI chat format to BetterChatGPT format
+export const convertOpenAIToBetterChatGPTFormat = (
+  openAIChat: OpenAIChat
+): ChatInterface => {
+  const messages: MessageInterface[] = [];
+
+  // Traverse the chat tree and collect messages
+  const traverseTree = (id: string) => {
+    const node = openAIChat.mapping[id];
+
+    // Extract message if it exists
+    if (node.message) {
+      const { role } = node.message.author;
+      const content = node.message.content.parts.join('');
+      if (content.length > 0) messages.push({ role, content });
+    }
+
+    // Traverse the last child node if any children exist
+    if (node.children.length > 0) {
+      traverseTree(node.children[node.children.length - 1]);
+    }
+  };
+
+  // Start traversing the tree from the root node
+  const rootNode = openAIChat.mapping[Object.keys(openAIChat.mapping)[0]].id;
+  traverseTree(rootNode);
+
+  // Return the chat interface object
+  return {
+    id: uuidv4(),
+    title: openAIChat.title,
+    messages,
+    config: _defaultChatConfig,
+    titleSet: true,
+  };
+};
+
+// Import OpenAI chat data and convert it to BetterChatGPT format
+export const importOpenAIChatExport = (openAIChatExport: OpenAIChat[]) => {
+  return openAIChatExport.map(convertOpenAIToBetterChatGPTFormat);
 };
